requirements.txt
# Python 3.10.12
tqdm==4.66.4
pandas==2.2.2
openai==1.30.1
jupyterlab==4.2.0
python-dotenv==1.0.1
streamlit~=1.34.0
deep_translator==1.9.1
requests~=2.31.0
pillow~=10.3.0